requirements.txt
# Core RAG and ML Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.22
openai>=1.12.0
cohere>=4.47

# Audio Processing and Speech Recognition
whisperx>=3.1.1
torch>=2.2.0
torchaudio>=2.2.0
transformers>=4.36.2
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# Data Processing and Analysis
pandas>=2.1.4
numpy>=1.24.3
scikit-learn>=1.3.2
nltk>=3.8.1
spacy>=3.7.2

# Web Framework and UI
streamlit>=1.29.0
streamlit-option-menu>=0.3.6
plotly>=5.17.0
altair>=5.2.0

# Vector Database and Search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Evaluation and Metrics
ragas>=0.1.0
evaluate>=0.4.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.1
Pillow>=10.1.0
matplotlib>=3.8.2
seaborn>=0.13.0

# Development and Testing
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0 